ml/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.6.0
python-dotenv==1.0.0
google-generativeai==0.3.1
nltk==3.8.1

